Add Chinese support
1. Add chinese translation of all strings: B9Creator_zh_CN.ts

Firstly, extract all of the strings which should be translated:

$ lupdate B9Creator.pro -ts B9Creator_zh_CN.ts

Second, translate the strings to chinese with the "linguist" tool, or simply
edit the .ts file ourselves. If finish, remove the "unfinished" type attribute.

Third, release the translation file and generate a compressed .qm file, which
can be used in the source code specify the release language.

2. Use our added chinese translation

To use our chinese translation, we must load the .qm translation file before
the real Windows launching.

We find we must added the translation codes in mainwindow.cpp, the main.cpp
doesn't work.

3. TODO

Only translated 6 strings, need to finish the others...

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>